refactor(core): move task DTOs into cuenv-manifest (Phase 2b)
Move the task type layer out of cuenv-core into the leaf DTO crate:
Task/TaskGroup/TaskNode/Tasks with their custom deserialization and
pure accessors, plus dependency, params, inputs, capture, cache
policy, retry, dagger, and shell DTO modules. Task-path resolution
(TaskResolver impl and path walking) moves too, so cuenv-manifest
gains a cuenv-task-graph dependency and owns the
TaskNodeData/MutableTaskNodeData impls (orphan rules).

cuenv-core keeps the execution side: command resolution becomes the
crate-private TaskCommandExt extension trait, EffectiveScriptShell/
TaskCommandSpec stay in a slimmed tasks/shell.rs, and the AffectedBy
impls remain with core's trait. The services and hook-item manifest
DTOs (which embed task types) migrate as well; core::tasks and
core::manifest re-export everything at the old paths.

Validation: workspace check --all-targets clean; tests green across
cuenv-manifest (18), cuenv-core lib (666 + known sandbox-only
process-tree failure), cuenv (433), cuenv-ci (334), cuenv-services
(45), task-discovery (28); clippy clean on touched crates under the
CI-pinned lint set (one struct_excessive_bools note from a newer
local clippy on the pre-existing ShellOptions schema DTO); cargo fmt;
git diff --check.
