+++ crates/core/src/tasks/tasks_tests.rs
@@ -1,4 +1,5 @@
 use super::*;
+use std::collections::HashMap;
 
 #[test]
 fn test_task_default_values() {
