+++ crates/manifest/src/tasks/dagger.rs
